refactor(observability): move ProjectTagFilter out of filter pkg to keep it SDK-free
ProjectTagFilter was the only symbol in pkg/observability/filter that
imported aws-sdk-go-v2/service/ec2/types, which made the whole filter
package transitively depend on the EC2 SDK. reliable imports
filter.{Project,EnsureProject} from its thin observability-via-oracle
proxy (reliable#2141 / #2153) and must not drag in any aws SDK service
package — its cmd/api Vercel function is at the 250 MB limit.

Move ProjectTagFilter (+ its test) into pkg/observability/discovery/aws,
whose inspectors are the only callers and already import the EC2 SDK.
filter is now SDK-free (go list -deps: 0 aws/gcp service packages),
unblocking the leaf-type cutover on the reliable side.

No behavior change: pure relocation, call sites updated to the local
identifier.
