+++ pkg/observability/filter/project_test.go
@@ -177,21 +177,6 @@ func TestEnsureProject(t *testing.T) {
 	}
 }
 
-func TestProjectTagFilter(t *testing.T) {
-	t.Parallel()
-	t.Run("empty project returns nil", func(t *testing.T) {
-		t.Parallel()
-		assert.Nil(t, ProjectTagFilter(""))
-	})
-	t.Run("non-empty project returns tag:Project filter", func(t *testing.T) {
-		t.Parallel()
-		filters := ProjectTagFilter("io-myproject")
-		assert.Len(t, filters, 1)
-		assert.Equal(t, "tag:Project", *filters[0].Name)
-		assert.Equal(t, []string{"io-myproject"}, filters[0].Values)
-	})
-}
-
 func TestMatch_KVFormat(t *testing.T) {
 	t.Parallel()
 	resources := []map[string]any{
